examples/gcp_functions/quotes/clean/requirements.txt
requests
google-cloud-pubsub
google-cloud-tasks
google-cloud-datastore